refactor: electron clipboard to common interface
