+++ packages/core-browser/src/bootstrap/inner-providers.ts
@@ -76,14 +76,7 @@ import { IProgressService } from '../progress';
 import { ProgressService } from '../progress/progress.service';
 import { AppConfig } from '../react-providers/config-provider';
 import { SlotRendererContribution } from '../react-providers/slot';
-import {
-  CredentialsService,
-  ICredentialsService,
-  CryptoService,
-  ICryptoService,
-  ElectronClipboardService,
-  INativeClipboardService,
-} from '../services';
+import { CredentialsService, ICredentialsService, CryptoService, ICryptoService } from '../services';
 import { IClipboardService, BrowserClipboardService } from '../services/clipboard.service';
 import { IExternalUriService, ExternalUriService } from '../services/external-uri.service';
 import { IToolbarPopoverRegistry, IToolbarRegistry, ToolBarActionContribution } from '../toolbar';
@@ -290,10 +283,6 @@ export function injectInnerProviders(injector: Injector) {
       {
         token: IElectronMenuBarService,
         useClass: ElectronMenuBarService,
-      },
-      {
-        token: INativeClipboardService,
-        useClass: ElectronClipboardService,
       },
     );
   }
