+++ packages/core-browser/src/services/index.ts
@@ -2,7 +2,6 @@ export * from './label-service';
 export * from './storage-service';
 export * from './status-bar-service';
 export * from './clipboard.service';
-export * from './electron-clipboard.service';
 export * from './external-uri.service';
 export * from './credentials-service';
 export * from './crypto-service';
