+++ packages/electron-basic/src/browser/clipboard/index.ts
