+++ packages/electron-basic/src/browser/index.ts
@@ -21,6 +21,7 @@ import {
   formatLocalize,
   IEventBus,
   Schemes,
+  IClipboardService,
 } from '@opensumi/ide-core-browser';
 import { ComponentContribution, ComponentRegistry } from '@opensumi/ide-core-browser/lib/layout';
 import { IMenuRegistry, MenuContribution, MenuId } from '@opensumi/ide-core-browser/lib/menu/next';
@@ -34,6 +35,7 @@ import {
 } from '@opensumi/ide-editor/lib/browser';
 import { IMessageService } from '@opensumi/ide-overlay/lib/common';
 
+import { ElectronClipboardService } from './clipboard';
 import { ElectronNativeDialogService } from './dialog';
 import { ElectronHeaderBar } from './header/header';
 import { WelcomeContribution } from './welcome/contribution';
@@ -47,6 +49,10 @@ export class ElectronBasicModule extends BrowserModule {
     },
     ElectronBasicContribution,
     WelcomeContribution,
+    {
+      token: IClipboardService,
+      useClass: ElectronClipboardService,
+    },
   ];
 }
 
@@ -371,6 +377,12 @@ export class ElectronBasicContribution
           }
         }
       }
+    });
+
+    // override broswer modules
+    this.injector.overrideProviders({
+      token: IClipboardService,
+      useClass: ElectronClipboardService,
     });
   }
 }
