+++ packages/file-tree-next/__tests__/browser/file-tree-contribution.test.ts
@@ -116,10 +116,6 @@ describe('FileTreeContribution', () => {
         useValue: {},
       },
       {
-        token: INativeClipboardService,
-        useValue: {},
-      },
-      {
         token: IDialogService,
         useValue: {},
       },
