+++ packages/file-tree-next/__tests__/browser/file-tree-model.service.test.ts
@@ -1,5 +1,4 @@
 import { TreeNodeType } from '@opensumi/ide-components';
-import { INativeClipboardService } from '@opensumi/ide-core-browser';
 import {
   URI,
   Disposable,
